django_rag_service/requirements.txt

requests==2.31.0
pandas==2.0.0
python-dotenv==1.0.0
langchain==0.3.7
langchain-community==0.3.5
openai==1.56.2
botbuilder-integration-aiohttp==4.14.7
pyadaptivecards==0.1.1
openpyxl==3.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
tabulate==0.9.0
django==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
adrf==0.1.4
gradio==4.13.0
streamlit==1.31.0
tiktoken==0.7.0
docx2txt==0.8
pgvector==0.2.5
pypdf==4.2.0

# For shell_plus in lower env's
django-extensions==3.2.3
langchain-openai==0.2.11

